[wpilibj] Move Preferences from util to preferences package
