+++ wpilibj/src/main/java/org/wpilib/preferences/Preferences.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.util;
+package org.wpilib.preferences;
 
 import static org.wpilib.util.ErrorMessages.requireNonNullParam;
