+++ wpilibjExamples/src/main/java/org/wpilib/examples/armsimulation/subsystems/Arm.java
@@ -10,6 +10,7 @@ import org.wpilib.hardware.rotation.Encoder;
 import org.wpilib.math.controller.PIDController;
 import org.wpilib.math.system.DCMotor;
 import org.wpilib.math.util.Units;
+import org.wpilib.preferences.Preferences;
 import org.wpilib.simulation.BatterySim;
 import org.wpilib.simulation.EncoderSim;
 import org.wpilib.simulation.RoboRioSim;
@@ -21,7 +22,6 @@ import org.wpilib.system.RobotController;
 import org.wpilib.telemetry.Telemetry;
 import org.wpilib.util.Color;
 import org.wpilib.util.Color8Bit;
-import org.wpilib.util.Preferences;
 
 public class Arm implements AutoCloseable {
   // The P gain for the PID controller that drives this arm.
