+++ wpilibjExamples/src/test/java/org/wpilib/examples/armsimulation/ArmSimulationTest.java
@@ -15,13 +15,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.wpilib.hardware.hal.HAL;
 import org.wpilib.hardware.hal.RobotMode;
 import org.wpilib.math.util.Units;
+import org.wpilib.preferences.Preferences;
 import org.wpilib.simulation.DriverStationSim;
 import org.wpilib.simulation.EncoderSim;
 import org.wpilib.simulation.JoystickSim;
 import org.wpilib.simulation.PWMMotorControllerSim;
 import org.wpilib.simulation.RoboRioSim;
 import org.wpilib.simulation.SimHooks;
-import org.wpilib.util.Preferences;
 
 @ResourceLock("timing")
 class ArmSimulationTest {
